+++ server/config.js
@@ -1,0 +1,5 @@
+module.exports = {
+    database: 'mongodb://localhost:27017/amazono',
+    port: 3030,
+    secret: '87654321'
+};