+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+4/20/2018
+- [Added] server/config.js
+
 4/15/2018
 - 使用 Node.js 與 Angular 5，建立模仿Amazon的網站
 - 安裝 [nodemon](https://github.com/remy/nodemon/) 來啟動 node 程式，參考[介紹](https://oranwind.org/node-nodemon/)